move constants.py to parent directory for global use
