+++ MVL_AI_Classifier/constants.py
