+++ MVL_AI_Classifier/features/noise_residuals_pipeline.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.ndimage import gaussian_filter
 
 from features.base_processor import BasePreprocessor
-from features.constants import DEFAULT_EPSILON, PATCH_SIZE
+from MVL_AI_Classifier.constants import DEFAULT_EPSILON, PATCH_SIZE
 from features.rgb_normalization_pipeline import RGBNormalizationPreprocessor
